fix(axvisor): recover riscv guest memory faults
