+++ components/axcpu/Cargo.toml
@@ -22,9 +22,10 @@ license = "Apache-2.0"
 
 [features]
 default = []
+exception-table = []
 fp-simd = []
 tls = []
-uspace = []
+uspace = ["exception-table"]
 arm-el2 = ["ax-page-table-entry/arm-el2"]
 xuantie-c9xx = ["ax-page-table-entry/xuantie-c9xx"]
