+++ components/axcpu/src/aarch64/init.rs
@@ -113,9 +113,10 @@ pub unsafe fn init_mmu(root_paddr: PhysAddr) {
 /// In detail, it initializes the exception vector, and sets `TTBR0_EL1` to 0 to
 /// block low address access.
 pub fn init_trap() {
+    #[cfg(feature = "exception-table")]
+    crate::exception_table::init_exception_table();
     #[cfg(feature = "uspace")]
     {
-        crate::uspace_common::init_exception_table();
         CNTKCTL_EL1.modify(CNTKCTL_EL1::EL0VCTEN::TrappedNone + CNTKCTL_EL1::EL0PCTEN::TrappedNone);
         barrier::isb(barrier::SY);
     }
