+++ components/axcpu/src/loongarch64/init.rs
@@ -53,8 +53,8 @@ pub fn init_mmu(root_paddr: PhysAddr, phys_virt_offset: usize) {
 ///
 /// In detail, it initializes the exception vector on LoongArch64 platforms.
 pub fn init_trap() {
-    #[cfg(feature = "uspace")]
-    crate::uspace_common::init_exception_table();
+    #[cfg(feature = "exception-table")]
+    crate::exception_table::init_exception_table();
     unsafe {
         unsafe extern "C" {
             fn exception_entry_base();
