+++ components/axcpu/src/riscv/init.rs
@@ -4,8 +4,8 @@
 ///
 /// In detail, it initializes the trap vector on RISC-V platforms.
 pub fn init_trap() {
-    #[cfg(feature = "uspace")]
-    crate::uspace_common::init_exception_table();
+    #[cfg(feature = "exception-table")]
+    crate::exception_table::init_exception_table();
     unsafe extern "C" {
         fn trap_vector_base();
     }
