+++ os/arceos/modules/axhal/Cargo.toml
@@ -46,7 +46,13 @@ paging = [
 ]
 tls = ["ax-cpu/tls"]
 uspace = ["paging", "ax-cpu/uspace", "axplat-dyn?/uspace"]
-hv = ["paging", "ax-cpu/arm-el2", "ax-percpu/arm-el2", "axplat-dyn?/hv"]
+hv = [
+    "paging",
+    "ax-cpu/arm-el2",
+    "ax-cpu/exception-table",
+    "ax-percpu/arm-el2",
+    "axplat-dyn?/hv",
+]
 axvisor-linker = []
 
 starry-kcov = []
